duration and List done

+++ src/main/java/com/isdcm/streamingapp/servlets/ServletRegistroVid.java
@@ -29,7 +29,7 @@ public class ServletRegistroVid extends HttpServlet {
                     request.getParameter("title"),
                     request.getParameter("autor"),
                     formatter.parse(request.getParameter("fechaCreacion")),
-                    Integer.parseInt(request.getParameter("duracion")),
+                    Float.parseFloat(request.getParameter("duracion")),
                     0,
                     request.getParameter("descripcion"),
                     request.getParameter("formato"),
